Sandro shubitidze

+++ group2/app/Actions/team/deleteTeam.php
@@ -2,6 +2,7 @@
 
 namespace App\Actions\team;
 
+use App\Http\Requests\GetTeamRequest;
 use App\Http\Requests\DeleteTeamRequest;
 use App\Models\teams;
 use Illuminate\Http\JsonResponse;
@@ -9,15 +10,16 @@ use Lorisleiva\Actions\Action;
 
 class deleteTeam extends Action
 {
-    public function handle(array $data)
-    {
-        $team = teams::find($data['id']);
+    public function handle(int $id): JsonResponse{
+        $team = teams::find($id);
+        if (!$team) {
+            return response()->json(['error' => 'Team not found'], 404);
+        }
         $team->delete();
-        return response() -> json(['message'=>'team deleted'], 200);
+        return response()->json(['message' => 'Team record deleted successfully'],200);
     }
 
-    public function asController(DeleteTeamRequest $request): JsonResponse{
-        return $this->handle($request->validated());
+    public function asController(int $id): JsonResponse {
+        return $this->handle($id);
     }
-
 }
